Split `test.rs` into modules
